Extract superclass implementation for packets
